+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.5] - 2020-09-12
+### Fixed
+- Filter out invalid characters from html ids
+### Updated
+- Cleanup timetable map and system map js
+
 ## [1.0.4] - 2020-09-12
 ### Fixed
 - Fix for showOnlyTimepoint filter